Add multiplexing capability in combination with the tca9548a
The library can now use any multiplexer that works in the same way as
the tca9548a device. An example on how to use the new functionality was
added under examples/distance_multiplexer.py. The code was tested on one
tca9548a device with three vl53l1x devices connected.
	geändert:       api/platform/vl53l1_platform.c
	geändert:       api/platform/vl53l1_platform.h
	geändert:       api/platform/vl53l1_platform_user_data.h
	neue Datei:     examples/distance_multiplexer.py
	geändert:       python/VL53L1X.py
	geändert:       python_lib/vl53l1x_python.c